update kirin_ws dsl structure

+++ src/tester/parse_dsl.py
@@ -93,7 +93,6 @@ def preprocess_dsl(
         logger.info("Single node detected.")
 
     sub_dsl_result = []
-    transformed_dsl_list = []
     for i, node_dsl in enumerate(node_dsl_list):
         i += 1
         logger.info(f"[#{i}] Node DSL parsing starts...")
@@ -109,7 +108,6 @@ def preprocess_dsl(
             split_not_has=split_not_has,
         )
         transformed_dsl = not_visitor.visit(node_tree)
-        transformed_dsl_list.append(transformed_dsl)
         logger.info(f"[#{i}] Node DSL transformation done~")
 
         # DSL decomposition -- starting from node Stmt -- KirinOrVisitor
@@ -133,7 +131,7 @@ def preprocess_dsl(
                 sub_dsl_result[i][j] = KirinRunner.format_dsl_text(sub_dsl)
 
     return DslPrepResDict(
-        node_dsl_list=node_dsl_list if not init_transform else transformed_dsl_list,
+        node_dsl_list=node_dsl_list,
         sub_dsl_collection=sub_dsl_result,
     )
 
@@ -146,19 +144,20 @@ def save_dsl_prep_res(dsl_prep_res: DslPrepResDict, dsl_dir: Path, is_opposite: 
     :param is_opposite: whether the dsl is in the opposite setting
     :return: None, the dsl preprocess result will be saved to dsl_dir
     """
-    assert dsl_dir.parent.is_dir(), f"--> DSL directory {dsl_dir.parent} not found!"
-    create_dir_with_path(dsl_dir, cleanup=True)
+    assert dsl_dir.is_dir(), f"--> DSL directory {dsl_dir} not found!"
 
     prefix = "DSL_OPP" if is_opposite else "DSL"
 
     for i, node_dsl in enumerate(dsl_prep_res["node_dsl_list"]):
         i += 1
-        node_dsl_path = dsl_dir / f"{prefix}_N{i}.kirin"
-        node_dsl_path.write_text(node_dsl, encoding="utf-8")
+        node_dsl_path = dsl_dir / f"DSL_N{i}.kirin"
+        if not node_dsl_path.exists():
+            node_dsl_path.write_text(node_dsl, encoding="utf-8")
 
     for i, sub_dsl_list in enumerate(dsl_prep_res["sub_dsl_collection"]):
         i += 1
-        sub_dsl_dir = dsl_dir / f"sub_n{i}"
+        sub_folder = "opp" if is_opposite else "norm"
+        sub_dsl_dir = dsl_dir / f"sub_n{i}" / sub_folder
         sub_dsl_dir.mkdir(parents=True, exist_ok=True)
         for j, sub_dsl in enumerate(sub_dsl_list):
             j += 1
@@ -170,14 +169,16 @@ def save_dsl_prep_res(dsl_prep_res: DslPrepResDict, dsl_dir: Path, is_opposite: 
 
 if __name__ == "__main__":
     # Test the parse_dsl function
-    dsl_path = Path("data/tmp/dsl_6.kirin")
+    dsl_path = Path("kirin_ws/ONLINE_Use_Unsafe_Algorithm_IDEA/dsl/DSL_ORI.kirin")
     dsl_text = dsl_path.read_text(encoding="utf-8")
     dsl_prep_res = preprocess_dsl(dsl_text, init_transform=True, spec_na_strategy=True, do_format=True)
 
     result_str = "==> Preprocess Result:\n"
 
-    for i, sub_dsl_list in enumerate(dsl_prep_res["sub_dsl_collection"]):
-        # fmt_sub_dsl_list = list(map(KirinRunner.format_dsl_text, sub_dsl_list))
-        result_str += f"--> Node {i + 1}:"
-        result_str += "\n" + f"\n{'-'*20}\n".join(sub_dsl_list)
-    logger.info(result_str)
+    # for i, sub_dsl_list in enumerate(dsl_prep_res["sub_dsl_collection"]):
+    #     # fmt_sub_dsl_list = list(map(KirinRunner.format_dsl_text, sub_dsl_list))
+    #     result_str += f"--> Node {i + 1}:"
+    #     result_str += "\n" + f"\n{'-'*20}\n".join(sub_dsl_list)
+    # logger.info(result_str)
+
+    save_dsl_prep_res(dsl_prep_res, dsl_path.parent)
